Remove MainScreenViewModelOutlets protocol from VM file
