+++ TransartistryApp/TransartistryApp/PresentationLayer/Modules/MainScreenModule/ViewModel/MainScreenViewModel.swift
@@ -2,14 +2,6 @@ import RxSwift
 import RxCocoa
 import RxRelay
 
-protocol MainScreenViewModelOutlets: ViewModelOutlets {
-    var cameraPickerDriver: Driver<Void> { get }
-    var photoPickerDriver: Driver<Void> { get }
-    
-    func cameraPickerButtonTapped()
-    func photoPickerButtonTapped()
-}
-
 final class MainScreenViewModel: BaseViewModel,
                                  MainScreenViewModelOutlets {
