+++ TransartistryApp/TransartistryApp/PresentationLayer/Modules/MainScreenModule/ViewModel/MainScreenViewModelOutlets.swift
@@ -1,0 +1,9 @@
+import RxCocoa
+
+protocol MainScreenViewModelOutlets: ViewModelOutlets {
+    var cameraPickerDriver: Driver<Void> { get }
+    var photoPickerDriver: Driver<Void> { get }
+    
+    func cameraPickerButtonTapped()
+    func photoPickerButtonTapped()
+}
